+++ libraries/units.rb
@@ -2,6 +2,14 @@
 # Cookbook Name:: systemd
 # Library:: Chef::Resource::SystemdAutomount
 # Library:: Chef::Resource::SystemdDevice
+# Library:: Chef::Resource::SystemdMount
+# Library:: Chef::Resource::SystemdPath
+# Library:: Chef::Resource::SystemdService
+# Library:: Chef::Resource::SystemdSlice
+# Library:: Chef::Resource::SystemdSocket
+# Library:: Chef::Resource::SystemdSwap
+# Library:: Chef::Resource::SystemdTarget
+# Library:: Chef::Resource.::SystemdTimer
 #
 # Copyright 2015 The Authors
 #
@@ -20,6 +28,14 @@
 
 require_relative 'systemd_automount'
 require_relative 'systemd_device'
+require_relative 'systemd_mount'
+require_relative 'systemd_path'
+require_relative 'systemd_service'
+require_relative 'systemd_slice'
+require_relative 'systemd_socket'
+require_relative 'systemd_swap'
+require_relative 'systemd_target'
+require_relative 'systemd_timer'
 require_relative 'unit'
 
 class Chef::Resource
@@ -37,9 +53,7 @@ class Chef::Resource
       yield
     end
   end
-end
 
-class Chef::Resource
   class SystemdDevice < Chef::Resource::SystemdUnit
     self.resource_name = :systemd_device
     provides :systemd_device
@@ -48,5 +62,114 @@ class Chef::Resource
       :device
     end
   end
-end
 
+  class SystemdMount < Chef::Resource::SystemdUnit
+    self.resource_name = :systemd_mount
+    provides :systemd_mount
+
+    def conf_type(_ = nil)
+      :mount
+    end
+
+    option_attributes Systemd::Mount::OPTIONS
+
+    def mount
+      yield
+    end
+  end
+
+  class SystemdPath < Chef::Resource::SystemdUnit
+    self.resource_name = :systemd_path
+    provides :systemd_path
+
+    def conf_type(_ = nil)
+      :path
+    end
+
+    option_attributes Systemd::Path::OPTIONS
+
+    def path
+      yield
+    end
+  end
+
+  class SystemdService < Chef::Resource::SystemdUnit
+    self.resource_name = :systemd_service
+    provides :systemd_service
+
+    def conf_type(_ = nil)
+      :service
+    end
+
+    option_attributes Systemd::Service::OPTIONS
+
+    def service
+      yield
+    end
+  end
+
+  class SystemdSlice < Chef::Resource::SystemdUnit
+    self.resource_name = :systemd_slice
+    provides :systemd_slice
+
+    def conf_type(_ = nil)
+      :slice
+    end
+
+    option_attributes Systemd::Slice::OPTIONS
+  end
+
+  class SystemdSocket < Chef::Resource::SystemdUnit
+    self.resource_name = :systemd_socket
+    provides :systemd_socket
+
+    def conf_type(_ = nil)
+      :socket
+    end
+
+    option_attributes Systemd::Socket::OPTIONS
+
+    def socket
+      yield
+    end
+  end
+
+  class SystemdSwap < Chef::Resource::SystemdUnit
+    self.resource_name = :systemd_swap
+    provides :systemd_swap
+
+    def conf_type(_ = nil)
+      :swap
+    end
+
+    option_attributes Systemd::Swap::OPTIONS
+
+    def swap
+      yield
+    end
+  end
+
+  class SystemdTarget < Chef::Resource::SystemdUnit
+    self.resource_name = :systemd_target
+    provides :systemd_target
+
+    def conf_type(_ = nil)
+      :target
+    end
+  end
+
+  class SystemdTimer < Chef::Resource::SystemdUnit
+    self.resource_name = :systemd_timer
+    provides :systemd_timer
+
+    def conf_type(_ = nil)
+      :timer
+    end
+
+    option_attributes Systemd::Timer::OPTIONS
+
+    def timer
+      yield
+    end
+  end
+end
